+++ s4e-backend/src/main/java/pl/cyfronet/s4e/controller/InstitutionController.java
@@ -108,7 +108,7 @@ public class InstitutionController {
             return institutionService.getAll(BasicInstitutionResponse.class);
         }
         return institutionService.getUserInstitutionsBy(appUserDetails.getUsername(),
-                List.of(AppRole.GROUP_MEMBER.name()),
+                List.of(AppRole.INST_MEMBER.name()),
                 BasicInstitutionResponse.class);
     }