+++ s4e-backend/src/main/resources/db/migration/V51__remove_obsolete_app_roles_from_user_role_table.sql
@@ -1,0 +1,3 @@
+UPDATE user_role SET role = 'INST_MEMBER' WHERE role = 'GROUP_MEMBER';
+
+DELETE FROM user_role WHERE role = 'INST_MANAGER' OR role = 'GROUP_MANAGER';